add user account management, registration and settings

+++ app/controllers/api/v1/admin/users_controller.rb
@@ -2,7 +2,7 @@
 
 class Api::V1::Admin::UsersController < Api::V1::AdminController
   before_action :require_super_admin!
-  before_action :set_user, only: [:update, :destroy]
+  before_action :set_user, only: [ :update, :destroy ]
 
   def index
     users = User.order(created_at: :desc)
@@ -20,10 +20,16 @@ class Api::V1::Admin::UsersController < Api::V1::AdminController
   end
 
   def create
-    user = User.new(user_params)
+    user = User.new(invitation_params)
 
     if user.save
-      render json: UserBlueprint.render_as_hash(user), status: :created
+      code = RegistrationCodeService.issue(user)
+
+      render json: {
+        user: UserBlueprint.render_as_hash(user),
+        registration_code: code,
+        registration_url: api_v1_admin_auth_registration_url(code: code)
+      }, status: :created
     else
       render json: { errors: user.errors.full_messages }, status: :unprocessable_entity
     end
@@ -38,6 +44,10 @@ class Api::V1::Admin::UsersController < Api::V1::AdminController
   end
 
   def destroy
+    if @user == current_user
+      return render json: { error: "You cannot delete your own account" }, status: :unprocessable_entity
+    end
+
     if @user.destroy
       render json: { message: "User deleted" }
     else
@@ -53,12 +63,17 @@ class Api::V1::Admin::UsersController < Api::V1::AdminController
 
   def user_params
     params.require(:user).permit(
-      :name,
+      :first_name,
+      :last_name,
       :email,
       :username,
       :password,
       :password_confirmation,
       :role
     )
+  end
+
+  def invitation_params
+    params.require(:user).permit(:email, :first_name, :last_name, :role)
   end
 end
